\docType{package}
\name{hiAnalyzer}
\alias{hiAnalyzer}
\alias{hiAnalyzer-package}
\title{Functions for analyzing and visualizing viral integration site data.}
\description{
  hiAnalyzer contains set of functions which allow users to
  analyze and visualize viral integration site data. Most
  of the data used by this package is an output of either
  hiAnnotator or hiReadsProcessor package.
}

